Update CANCMDDC.ino
byte index for int index in addSessionConsist

+++ CANCMDDC/CANCMDDC.ino
@@ -2293,7 +2293,7 @@ void addSessionConsist(byte session, byte consist)
 #endif
 
   // does the session belong to this controller?
-  int index = getSessionIndex(session);
+  byte index = getSessionIndex(session);
 
   if (index == SF_UNHANDLED)
     return;
@@ -2304,6 +2304,7 @@ void addSessionConsist(byte session, byte consist)
   // add the consist address to the loco
   // remove for now -
   //invalid narrowing conversion from "int" to "unsigned char"
+
   //controllers[index].consist = { (consist & 0x7f), 0, ((consist & 0x80) == 0x80) };
 }
